Added initial support for payment submethods
This patch adds the option for the value returned by a payment
controller: if may return not only a single description of a payment
method but also an array of such method descriptions. Thus, a number
of different payment methods can be implemented by a single payment
module (same as in shipping modules).

Signed-off-by: Pavel Avrorov <avrorov@priboy.online>